context_aware_ocr/requirements.txt
streamlit
transformers
torch
timm
Pillow
requests
